Change setting up of mem object to allow a second reference to it.
Before, setting up a mem object created the mem table and the mem object, with a single column 'row_id'. Columns were added afterwards by calls to 'add_mem_col'

Now you can optionally pass the entire table definition as an argument. If present, the mem table is set up with all columns, and a mem object is created and returned. If omitted, it is assumed that the table exists and another reference to it is required. A new mem object is created from the existing table defintion and returned.